requirements.txt
scikit_learn
unidecode
nmslib-metabrainz
psycopg2-binary
Flask
tabulate
